refactor: move ic shared module to mgmt
