+++ src/console/src/factory/orbiter.rs
@@ -6,7 +6,7 @@ use candid::Principal;
 use junobuild_shared::constants::CREATE_ORBITER_CYCLES;
 use junobuild_shared::ic::create_canister_install_code;
 use junobuild_shared::mgmt::cmc::cmc_create_canister_install_code;
-use junobuild_shared::mgmt::types::SubnetId;
+use junobuild_shared::mgmt::types::cmc::SubnetId;
 use junobuild_shared::types::interface::CreateCanisterArgs;
 use junobuild_shared::types::state::{MissionControlId, UserId};
