+++ src/libs/shared/src/mgmt/impls.rs
@@ -1,4 +1,4 @@
-use crate::mgmt::types::CreateCanisterError;
+use crate::mgmt::types::cmc::CreateCanisterError;
 
 impl std::fmt::Display for CreateCanisterError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
